chore: bootstrap release-please (#2)
* chore: bootstrap release-please

* feat(ci): add GitHub workflows for commit linting and releases

Introduce commit linting and release automation workflows to ensure 
consistent commit messages and streamline release processes. The 
commitlint workflow validates commit messages on pull requests, 
using a custom configuration to enforce maximum line lengths. The 
release-please workflow manages versioning and releases when 
pushing to the main branch, requiring a PAT for authentication. 
Additionally, a build workflow is added for running plugin linter 
and tester in isolated environments.

* feat: migrate release-please configuration to new file

Move release-please configuration from workflow file to a dedicated 
configuration file. This change simplifies the workflow and improves 
maintainability by separating the configuration from the execution 
logic. Additionally, it enables the handling of GitHub releases and 
the use of a manifest file for better release management.

* chore: standardize workflow names and add pre-commit checks

Rename the build workflow to lowercase for consistency. Introduce a new
pre-commit workflow to ensure code quality by running pre-commit hooks 
on pull requests and pushes to the main branch. This enhances the
development process by enforcing coding standards automatically.

* fix: add missing newlines in configuration files

Ensure that the configuration files contain proper newlines at the end. 
This improves file consistency and adheres to common coding standards. 
Update the .gitignore file to remove an unnecessary blank line.

* chore: update pre-commit-hooks to version 5.0.0

Upgrade the pre-commit-hooks dependency to version 5.0.0 to ensure 
access to the latest features and fixes, improving the quality 
and consistency of coding standards in the project.

* chore: add .editorconfig and update pre-commit config

Add an .editorconfig file to standardize coding styles across the
project by defining rules for indentation, end of line characters,
and whitespace trimming. Update pre-commit configuration to ensure
that the shfmt hook uses an indentation of 2 spaces, aligning with
the new editorconfig settings. These changes enhance code 
consistency and collaboration.